mon_bot_heroku/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
attrs==25.3.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
enum-tools==0.13.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
six==1.17.0
twitchAPI==4.5.0
typing_extensions==4.14.0
uritemplate==4.1.1
urllib3==2.4.0
yarl==1.20.0
Flask